Chapter08/Recipe1/cdk-deployment-recipe/requirements.txt
aws-cdk-lib==2.106.1
constructs>=10.0.0,<11.0.0
git-remote-codecommit
